9.5

+++ typo3/sysext/install/Classes/Service/LanguagePackService.php
@@ -16,10 +16,13 @@ namespace TYPO3\CMS\Install\Service;
  */
 
 use Symfony\Component\Finder\Finder;
+use TYPO3\CMS\Core\Configuration\Features;
 use TYPO3\CMS\Core\Core\Environment;
+use TYPO3\CMS\Core\Http\RequestFactory;
 use TYPO3\CMS\Core\Localization\Locales;
 use TYPO3\CMS\Core\Package\PackageManager;
 use TYPO3\CMS\Core\Registry;
+use TYPO3\CMS\Core\Service\Archive\ZipService;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Utility\PathUtility;
@@ -43,10 +46,27 @@ class LanguagePackService
      */
     protected $registry;
 
+    /**
+     * @var RequestFactory
+     */
+    protected $requestFactory;
+
+    private const OLD_LANGUAGE_PACK_URLS = [
+        'https://extensions.typo3.org/fileadmin/l10n/',
+        'https://typo3.org/fileadmin/ter/',
+        'https://beta-translation.typo3.org/fileadmin/ter/',
+        'https://localize.typo3.org/fileadmin/ter/'
+    ];
+
+    private const OLD_LANGUAGE_PACK_URL = 'https://extensions.typo3.org/fileadmin/l10n/';
+
+    private const LANGUAGE_PACK_URL = 'https://localize.typo3.org/xliff/';
+
     public function __construct()
     {
         $this->locales = GeneralUtility::makeInstance(Locales::class);
         $this->registry = GeneralUtility::makeInstance(Registry::class);
+        $this->requestFactory = GeneralUtility::makeInstance(RequestFactory::class);
     }
 
     /**
@@ -66,7 +86,8 @@ class LanguagePackService
      */
     public function getActiveLanguages(): array
     {
-        return $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['lang']['availableLanguages'] ?? [];
+        $availableLanguages = $GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['lang']['availableLanguages'] ?? [];
+        return array_filter($availableLanguages);
     }
 
     /**
@@ -139,8 +160,10 @@ class LanguagePackService
             $extension = [
                 'key' => $key,
                 'title' => $title,
-                'icon' => PathUtility::stripPathSitePrefix(ExtensionManagementUtility::getExtensionIcon($path, true)),
             ];
+            if (!empty(ExtensionManagementUtility::getExtensionIcon($path, false))) {
+                $extension['icon'] = PathUtility::stripPathSitePrefix(ExtensionManagementUtility::getExtensionIcon($path, true));
+            }
             $extension['packs'] = [];
             foreach ($activeLanguages as $iso) {
                 $isLanguagePackDownloaded = is_dir(Environment::getLabelsPath() . '/' . $iso . '/' . $key . '/');
@@ -171,19 +194,22 @@ class LanguagePackService
      */
     public function updateMirrorBaseUrl(): string
     {
+        $repositoryUrl = 'https://repositories.typo3.org/mirrors.xml.gz';
         $downloadBaseUrl = false;
         try {
-            $xmlContent = GeneralUtility::getUrl('https://repositories.typo3.org/mirrors.xml.gz');
-            $xmlContent = GeneralUtility::xml2array(@gzdecode($xmlContent));
-            if (!empty($xmlContent['mirror']['host']) && !empty($xmlContent['mirror']['path'])) {
-                $downloadBaseUrl = 'https://' . $xmlContent['mirror']['host'] . $xmlContent['mirror']['path'];
+            $response = $this->requestFactory->request($repositoryUrl);
+            if ($response->getStatusCode() === 200) {
+                $xmlContent = @gzdecode($response->getBody()->getContents());
+                if (!empty($xmlContent['mirror']['host']) && !empty($xmlContent['mirror']['path'])) {
+                    $downloadBaseUrl = 'https://' . $xmlContent['mirror']['host'] . $xmlContent['mirror']['path'];
+                }
             }
         } catch (\Exception $e) {
             // Catch generic exception, fallback handled below
         }
         if (empty($downloadBaseUrl)) {
             // Hard coded fallback if something went wrong fetching & parsing mirror list
-            $downloadBaseUrl = 'https://typo3.org/fileadmin/ter/';
+            $downloadBaseUrl = self::OLD_LANGUAGE_PACK_URL;
         }
         $this->registry->set('languagePacks', 'baseUrl', $downloadBaseUrl);
         return $downloadBaseUrl;
@@ -222,6 +248,12 @@ class LanguagePackService
         if (empty($languagePackBaseUrl)) {
             throw new \RuntimeException('Language pack baseUrl not found', 1520169691);
         }
+
+        if (in_array($languagePackBaseUrl, self::OLD_LANGUAGE_PACK_URLS, true)
+            && GeneralUtility::makeInstance(Features::class)->isFeatureEnabled('newTranslationServer')) {
+            $languagePackBaseUrl = self::LANGUAGE_PACK_URL;
+        }
+
         // Allow to modify the base url on the fly by calling a signal
         $signalSlotDispatcher = GeneralUtility::makeInstance(Dispatcher::class);
         $signalSlotDispatcher->dispatch(
@@ -237,11 +269,11 @@ class LanguagePackService
         $majorVersion = explode('.', TYPO3_branch)[0];
         if (strpos($path, '/sysext/') !== false) {
             // This is a system extension and the package URL should be adapted to have different packs per core major version
-            // https://typo3.org/fileadmin/ter/b/a/backend-l10n/backend-l10n-fr.v9.zip
+            // https://extensions.typo3.org/fileadmin/l10n/b/a/backend-l10n/backend-l10n-fr.v9.zip
             $packageUrl = $key[0] . '/' . $key[1] . '/' . $key . '-l10n/' . $key . '-l10n-' . $iso . '.v' . $majorVersion . '.zip';
         } else {
             // Typical non sysext path, Hungarian:
-            // https://typo3.org/fileadmin/ter/a/n/anextension-l10n/anextension-l10n-hu.zip
+            // https://extensions.typo3.org/fileadmin/l10n/a/n/anextension-l10n/anextension-l10n-hu.zip
             $packageUrl = $key[0] . '/' . $key[1] . '/' . $key . '-l10n/' . $key . '-l10n-' . $iso . '.zip';
         }
 
@@ -255,19 +287,22 @@ class LanguagePackService
 
         $operationResult = false;
         try {
-            $languagePackContent = GeneralUtility::getUrl($languagePackBaseUrl . $packageUrl);
-            if (!empty($languagePackContent)) {
-                $operationResult = true;
-                if ($packExists) {
-                    $operationResult = GeneralUtility::rmdir($absoluteExtractionPath, true);
-                }
-                if ($operationResult) {
-                    GeneralUtility::mkdir_deep(Environment::getVarPath() . '/transient/');
-                    $operationResult = GeneralUtility::writeFileToTypo3tempDir($absolutePathToZipFile, $languagePackContent) === null;
-                }
-                $this->unzipTranslationFile($absolutePathToZipFile, $absoluteLanguagePath);
-                if ($operationResult) {
-                    $operationResult = unlink($absolutePathToZipFile);
+            $response = $this->requestFactory->request($languagePackBaseUrl . $packageUrl);
+            if ($response->getStatusCode() === 200) {
+                $languagePackContent = $response->getBody()->getContents();
+                if (!empty($languagePackContent)) {
+                    $operationResult = true;
+                    if ($packExists) {
+                        $operationResult = GeneralUtility::rmdir($absoluteExtractionPath, true);
+                    }
+                    if ($operationResult) {
+                        GeneralUtility::mkdir_deep(Environment::getVarPath() . '/transient/');
+                        $operationResult = GeneralUtility::writeFileToTypo3tempDir($absolutePathToZipFile, $languagePackContent) === null;
+                    }
+                    $this->unzipTranslationFile($absolutePathToZipFile, $absoluteLanguagePath);
+                    if ($operationResult) {
+                        $operationResult = unlink($absolutePathToZipFile);
+                    }
                 }
             }
         } catch (\Exception $e) {
@@ -319,43 +354,17 @@ class LanguagePackService
      *
      * @param string $file path to zip file
      * @param string $path path to extract to
-     * @throws \RuntimeException
      */
     protected function unzipTranslationFile(string $file, string $path)
     {
-        $zip = zip_open($file);
-        if (is_resource($zip)) {
-            if (!is_dir($path)) {
-                GeneralUtility::mkdir_deep($path);
-            }
-            while (($zipEntry = zip_read($zip)) !== false) {
-                $zipEntryName = zip_entry_name($zipEntry);
-                if (strpos($zipEntryName, '/') !== false) {
-                    $zipEntryPathSegments = explode('/', $zipEntryName);
-                    $fileName = array_pop($zipEntryPathSegments);
-                    // It is a folder, because the last segment is empty, let's create it
-                    if (empty($fileName)) {
-                        GeneralUtility::mkdir_deep($path . implode('/', $zipEntryPathSegments));
-                    } else {
-                        $absoluteTargetPath = GeneralUtility::getFileAbsFileName($path . implode('/', $zipEntryPathSegments) . '/' . $fileName);
-                        if (trim($absoluteTargetPath) !== '') {
-                            $return = GeneralUtility::writeFile(
-                                $absoluteTargetPath,
-                                zip_entry_read($zipEntry, zip_entry_filesize($zipEntry))
-                            );
-                            if ($return === false) {
-                                throw new \RuntimeException('Could not write file ' . $zipEntryName, 1520170845);
-                            }
-                        } else {
-                            throw new \RuntimeException('Could not write file ' . $zipEntryName, 1520170846);
-                        }
-                    }
-                } else {
-                    throw new \RuntimeException('Extension directory missing in zip file!', 1520170847);
-                }
-            }
-        } else {
-            throw new \RuntimeException('Unable to open zip file ' . $file, 1520170848);
+        if (!is_dir($path)) {
+            GeneralUtility::mkdir_deep($path);
         }
+
+        $zipService = GeneralUtility::makeInstance(ZipService::class);
+        if ($zipService->verify($file)) {
+            $zipService->extract($file, $path);
+        }
+        GeneralUtility::fixPermissions($path, true);
     }
 }
